backend/requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
azure-core==1.30.2
azure-iot-hub==2.6.1
billiard==4.2.1
Brotli==1.1.0
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap5==2025.4
cron-descriptor==1.4.5
cryptography==45.0.3
cssselect2==0.8.0
daphne==4.2.0
diff-match-patch==20241021
Django==5.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-cron==0.6.0
django-decouple==2.1
django-dynamic-formsets==0.0.8
django-filter==25.1
django-formtools==2.5.1
django-import-export==4.3.5
django-jazzmin==3.0.1
django-multiselectfield==0.1.13
django-restframework==0.0.1
django-timezone-field==7.1
django-unfold==0.59.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
firebase-admin==6.9.0
fonttools==4.58.0
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isodate==0.6.1
kombu==5.5.4
msgpack==1.1.0
msrest==0.7.1
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyphen==0.17.2
python-barcode==0.15.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
razorpay==1.4.2
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
uamqp==1.6.9
uritemplate==4.2.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
zope.interface==7.2
zopfli==0.2.3.post1
